+++ src/components/trackedEntity/TrackedEntityTypeSelect.js
@@ -1,55 +1,52 @@
+import { useDataQuery } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from 'prop-types'
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import { loadTrackedEntityTypes } from '../../actions/trackedEntities.js'
+import React from 'react'
 import { SelectField } from '../core/index.js'
 
-class TrackedEntityTypeSelect extends Component {
-    static propTypes = {
-        loadTrackedEntityTypes: PropTypes.func.isRequired,
-        onChange: PropTypes.func.isRequired,
-        className: PropTypes.string,
-        errorText: PropTypes.string,
-        trackedEntityType: PropTypes.object,
-        trackedEntityTypes: PropTypes.array,
-    }
-
-    componentDidMount() {
-        const { trackedEntityTypes, loadTrackedEntityTypes } = this.props
-
-        if (!trackedEntityTypes) {
-            loadTrackedEntityTypes()
-        }
-    }
-
-    render() {
-        const {
-            trackedEntityType,
-            trackedEntityTypes,
-            onChange,
-            className,
-            errorText,
-        } = this.props
-
-        return (
-            <SelectField
-                label={i18n.t('Tracked Entity Type')}
-                loading={trackedEntityTypes ? false : true}
-                items={trackedEntityTypes}
-                value={trackedEntityType ? trackedEntityType.id : null}
-                onChange={(trackedEntityType) => onChange(trackedEntityType)}
-                className={className}
-                errorText={!trackedEntityType && errorText ? errorText : null}
-                dataTest="tetypeselect"
-            />
-        )
-    }
+// Load all tracked entity types
+const TRACKED_ENTITY_TYPES_QUERY = {
+    types: {
+        resource: 'trackedEntityTypes',
+        params: {
+            fields: ['id', 'displayName~rename(name)'],
+            paging: false,
+        },
+    },
 }
 
-export default connect(
-    (state) => ({
-        trackedEntityTypes: state.trackedEntityTypes,
+const TrackedEntityTypeSelect = ({
+    trackedEntityType,
+    onChange,
+    className,
+    errorText,
+}) => {
+    const { loading, error, data } = useDataQuery(TRACKED_ENTITY_TYPES_QUERY)
+
+    return (
+        <SelectField
+            label={i18n.t('Tracked Entity Type')}
+            loading={loading}
+            items={data?.types.trackedEntityTypes}
+            value={trackedEntityType?.id}
+            onChange={onChange}
+            className={className}
+            errorText={
+                error?.message ||
+                (!trackedEntityType && errorText ? errorText : null)
+            }
+            dataTest="tetypeselect"
+        />
+    )
+}
+
+TrackedEntityTypeSelect.propTypes = {
+    onChange: PropTypes.func.isRequired,
+    className: PropTypes.string,
+    errorText: PropTypes.string,
+    trackedEntityType: PropTypes.shape({
+        id: PropTypes.string.isRequired,
     }),
-    { loadTrackedEntityTypes }
-)(TrackedEntityTypeSelect)
+}
+
+export default TrackedEntityTypeSelect